requirements.txt
django
python-social-auth
python-slugify
django-uuslug
pillow
pymongo
south
psycopg2
reportlab==2.6
xhtml2pdf
django-rosetta
djangorestframework